Several months of work
compiling of hundreds of data points for columns and beams. spredsheets to find CSM loads and Eurocode loads and curvatures and strains. work on curvature and written thesis. Updated compression code to retrieve datat from spreadsheet and wrote code to write results directly on spreadsheet . code to calculate effect of residual stresses. code to experiment with number of strips. and so much more that ive forgotten